+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import Router from './components/Router'
+import Router from 'components/Router'
 import CssBaseline from '@material-ui/core/CssBaseline'
-import ThemeProvider from './components/ThemeProvider'
+import ThemeProvider from 'context/ThemeProvider'
 
 function App () {
   return (
