+++ src/components/Router.scss
@@ -1,4 +1,4 @@
-@import '../constants/index.scss';
+@import '~constants/index.scss';
 
 .content {
     padding-top: 14px;
