+++ src/components/alerts/RegisterAlert.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 import Alert from '@material-ui/lab/Alert'
-import './index.css'
+import './RegisterAlert.scss'
 
 export default function RegisterAlert () {
   return (
