+++ src/components/alerts/RegisterAlert.scss
