+++ src/components/buttons/RadioGroupButtons.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react'
 import { TextField } from '@material-ui/core'
-import Radio from '@material-ui/core/Radio'
-import RadioGroup from '@material-ui/core/RadioGroup'
 import FormControlLabel from '@material-ui/core/FormControlLabel'
 import FormHelperText from '@material-ui/core/FormHelperText'
 import FormLabel from '@material-ui/core/FormLabel'
+import Radio from '@material-ui/core/Radio'
+import RadioGroup from '@material-ui/core/RadioGroup'
 
 export default function RadioGroupButtons (props) {
   const [otherDisabled, disableOther] = useState(true)
