+++ src/components/layout/Navigation.scss
@@ -1,4 +1,4 @@
-@import '../constants/index.scss';
+@import '~constants/index.scss';
 
 .Navigation {
     height: 100%;
