+++ src/components/routing/Logout.js
@@ -1,8 +1,8 @@
 import React from 'react'
+import { connect } from 'react-redux'
 import { Auth } from 'aws-amplify'
 import Button from '@material-ui/core/Button'
-import { logout } from '../../actions/UserActions'
-import { connect } from 'react-redux'
+import { logout } from 'store/user/UserActions'
 
 export function Logout (props) {
   return (
