+++ src/pages/member/Event/QuickRegister.scss
@@ -1,4 +1,4 @@
-@import '../../../constants/index.scss';
+@import '~constants/index.scss';
 
 .layout {
     width: 66vw;
