+++ src/pages/public/Event/EventRegister.js
@@ -5,8 +5,8 @@ import { Helmet } from 'react-helmet'
 import { Formik } from 'formik'
 import * as Yup from 'yup'
 
-import EventView from 'components/EventView'
-import RegisterEvent from 'components/Forms/RegisterEvent'
+import EventView from 'components/Event/EventView'
+import EventRegisterForm from './EventRegisterForm'
 import NotFound from 'pages/NotFound'
 
 import { makeStyles } from '@material-ui/core/styles'
@@ -78,7 +78,7 @@ const EventFormContainer = (props) => {
               validationSchema={validationSchema}
               onSubmit={submitValues}
             >
-              {props => <RegisterEvent {...props} />}
+              {props => <EventRegisterForm {...props} />}
             </Formik>
           </EventView>
         </Paper>
