+++ src/store/rootStore.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux'
-import rootReducer from '../reducers/rootReducer'
+import rootReducer from './rootReducer'
 
 const store = createStore(rootReducer)
